pyproject.toml
[project]
name = "simple-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hubaib", email = "hubaibm9@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.9",
]

[project.scripts]
simple-mcp = "simple_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
